Changed type of slice width from float to double
This was in the hopes that it would make the parameter
parsable by the plugin stuff (which may not actually be a problem),
but it didn't seem to do anything.

+++ Plugins/EnsembleSurfaceSlicing/vtkEnsembleSurfaceSlicingRepresentation.h
@@ -29,8 +29,8 @@ public:
   static vtkEnsembleSurfaceSlicingRepresentation* New();
   vtkTypeMacro(vtkEnsembleSurfaceSlicingRepresentation, vtkGeometryRepresentationWithFaces);
 
-  void SetSliceWidth(float width);
-  vtkGetMacro(SliceWidth, float);
+  void SetSliceWidth(double width);
+  vtkGetMacro(SliceWidth, double);
 
 //BTX
 protected:
@@ -42,7 +42,7 @@ private:
   void operator=(const vtkEnsembleSurfaceSlicingRepresentation&); // Not implemented
 //ETX
 
-  float SliceWidth;
+  double SliceWidth;
 };
 
 #endif
